feat(frontend-auth): authentication update
